+++ packages/happn-cluster/test/mocks/mock-opts.js
@@ -2,7 +2,6 @@ module.exports = {
   logger: {
     createLogger: function () {
       return {
-        $$TRACE: function () {},
         debug: function () {
           if (process.env.LOG_LEVEL === 'off');
         },
@@ -17,6 +16,23 @@ module.exports = {
         },
         fatal: function () {
           if (process.env.LOG_LEVEL === 'off');
+        },
+        json: {
+          debug: function () {
+            if (process.env.LOG_LEVEL === 'off');
+          },
+          info: function () {
+            if (process.env.LOG_LEVEL === 'off');
+          },
+          warn: function () {
+            if (process.env.LOG_LEVEL === 'off');
+          },
+          error: function () {
+            if (process.env.LOG_LEVEL === 'off');
+          },
+          fatal: function () {
+            if (process.env.LOG_LEVEL === 'off');
+          },
         },
       };
     },