+++ packages/happner-cluster/test/integration/09-cluster-broker.js
@@ -1,4 +1,3 @@
-const Promise = require('bluebird');
 const libDir = require('../_lib/lib-dir');
 const baseConfig = require('../_lib/base-config');
 const stopCluster = require('../_lib/stop-cluster');
@@ -89,8 +88,7 @@ require('../_lib/test-helper').describe({ timeout: 120e3 }, (test) => {
   }
 
   beforeEach('clear mongo collection', function (done) {
-    stopCluster(servers, function (e) {
-      if (e) return done(e);
+    stopCluster(servers, function () {
       servers = [];
       clearMongoCollection('mongodb://localhost', 'happn-cluster', function () {
         done();
@@ -256,6 +254,11 @@ require('../_lib/test-helper').describe({ timeout: 120e3 }, (test) => {
           );
         })
         .then(function () {
+          return new Promise((resolve) => {
+            setTimeout(resolve, 3000);
+          });
+        })
+        .then(function () {
           return testclient.create('username', 'password', getSeq.getPort(2));
         })
         .then(function (client) {
@@ -366,6 +369,11 @@ require('../_lib/test-helper').describe({ timeout: 120e3 }, (test) => {
           return users.allowEvent(localInstance, 'username', 'remoteComponent', '/brokered/event');
         })
         .then(function () {
+          return new Promise(function (resolve) {
+            setTimeout(resolve, 3000);
+          });
+        })
+        .then(function () {
           return testclient.create('username', 'password', getSeq.getPort(2));
         })
         .then(function (client) {
@@ -469,6 +477,11 @@ require('../_lib/test-helper').describe({ timeout: 120e3 }, (test) => {
           );
         })
         .then(function () {
+          return new Promise(function (resolve) {
+            setTimeout(resolve, 3000);
+          });
+        })
+        .then(function () {
           return testclient.create('username', 'password', getSeq.getPort(2));
         })
         .then(function (client) {
@@ -490,6 +503,11 @@ require('../_lib/test-helper').describe({ timeout: 120e3 }, (test) => {
             'remoteComponent',
             'brokeredMethodFail'
           );
+        })
+        .then(function () {
+          return new Promise(function (resolve) {
+            setTimeout(resolve, 3000);
+          });
         })
         .then(function () {
           return testclient.create('username', 'password', getSeq.getPort(2));
@@ -517,6 +535,11 @@ require('../_lib/test-helper').describe({ timeout: 120e3 }, (test) => {
           );
         })
         .then(function () {
+          return new Promise(function (resolve) {
+            setTimeout(resolve, 3000);
+          });
+        })
+        .then(function () {
           return testclient.create('username', 'password', getSeq.getPort(2));
         })
         .then(function (client) {
@@ -532,21 +555,26 @@ require('../_lib/test-helper').describe({ timeout: 120e3 }, (test) => {
     it('ensures an error is handled and returned accordingly if we execute a method that does not exist on the cluster mesh yet', function (done) {
       startClusterEdgeFirst()
         .then(function () {
-          return users.allowMethod(localInstance, 'username', 'brokerComponent', 'directMethod');
+          return users.allowMethod(localInstance, 'username', 'brokerComponent10', 'directMethod');
         })
         .then(function () {
-          return users.allowMethod(localInstance, 'username', 'remoteComponent', 'brokeredMethod1');
+          return users.allowMethod(
+            localInstance,
+            'username',
+            'remoteComponent',
+            'brokeredMethod10'
+          );
         })
         .then(function () {
           return testclient.create('username', 'password', getSeq.getPort(1));
         })
         .then(function (client) {
-          return client.exchange.remoteComponent.brokeredMethod1();
+          return client.exchange.remoteComponent.brokeredMethod10();
         })
         .catch(function (e) {
           test
             .expect(e.toString())
-            .to.be('Error: Not implemented remoteComponent:^2.0.0:brokeredMethod1');
+            .to.be('Error: Not implemented remoteComponent:^2.0.0:brokeredMethod10');
           setTimeout(done, 2000);
         });
     });