+++ RELEASES.md
@@ -10,6 +10,16 @@ happner-2: 12.1.1
 happner-client: 12.2.0
 happner-cluster: 12.1.1
 
+2.0.0 2022-04-18
+-----------------
+happn-3: 13.0.7
+happn-cluster: 13.0.0
+happn-db-provider-elasticsearch: 1.0.5
+happn-db-provider-mongo: 1.0.5
+happner-client: 12.2.0
+happner-cluster: 13.0.0
+
+
 
 1.2.1 2022-05-04
 ----------------