+++ packages/happn-commons/RELEASES.md
@@ -2,3 +2,7 @@
 -----------------
   - TEN-92, TEN-93: logging levels and events
   - TEN-102: loki snapshot  file redundancy
+
+1.0.3 2022-05-17
+-----------------
+  - SMC-4817: mongo-based membership strategy: printOpenHandles