+++ packages/happn-db-provider-elasticsearch/RELEASES.md
@@ -32,3 +32,7 @@
 1.0.7 2022-05-09
 -----------------
   - TEN-114: happner-2: enforce MESH_GST, MESH_ADMIN upsert on startup, ensures db dir
+
+1.0.8 2022-05-17
+-----------------
+  - SMC-4817: mongo-based membership strategy